# src/main/resources/api.yaml
swagger: '2.0'
schemes:
  - http
host: todos.stoplight.io
info:
  version: ''
  title: demo rest api
  description: ''
paths:
  /persons:
    parameters: []
    get:
      tags:
        - Person
      summary: List Person
      description: ''
      operationId: GET_persons
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            type: array
            items:
              $ref: '#/definitions/person'
          examples:
            application/json:
              - id: 17
                name: new lit item
                completed: null
                completed_at: null
                url: 'https://todos.stoplight.io/todos/17'
                created_at: '2014-08-28T14:14:28.494Z'
  '/person/{personId}':
    parameters:
      - name: personId
        in: path
        required: true
        type: string
    get:
      tags:
        - Person
      summary: Get Person
      description: ''
      operationId: GET_person-personId
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/person'
          examples:
            application/json:
              id: 17
              name: new lit item
              completed: null
              completed_at: null
              created_at: '2014-08-28T14:14:28.494Z'
              updated_at: '2014-08-28T14:14:28.494Z'
    put:
      tags:
        - Person
      summary: Update Todo
      description: ''
      operationId: PUT_person-personId
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              completed:
                type: boolean
            required:
              - name
            example:
              name: my todo's new name
              completed: false
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/person'
        '403':
          description: ''
          schema:
            type: object
            properties: {}
    delete:
      tags:
        - Person
      summary: Delete Person
      description: ''
      operationId: DELETE_person-personId
      consumes:
        - application/json
      produces:
        - application/json
      parameters: []
      responses:
        '204':
          description: ''
          schema:
            type: object
            properties: {}
  /person:
    parameters: []
    post:
      tags:
        - Person
      summary: Create Person
      description: ''
      operationId: POST_person
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              completed:
                type: boolean
            required:
              - name
            example:
              name: my todo's name
              completed: false
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/person'
        '403':
          description: ''
          schema:
            type: object
            properties: {}
definitions:
  person:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      suname:
        type:
          - string
          - 'null'
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      employeeNo:
        type:
          - string
          - 'null'
      email:
        type: string
        format: email
      age:
        type: number
        minimum: 0
        maximum: 100
    required:
      - id
      - name
      - created_at